# docs/swagger.yaml
basePath: /
definitions:
  config.Credentials:
    properties:
      apikey:
        type: string
      password:
        type: string
      secret:
        type: string
      username:
        type: string
    type: object
  config.DataSource:
    properties:
      creds:
        additionalProperties:
          $ref: '#/definitions/config.Credentials'
        type: object
      name:
        type: string
      ttl:
        type: integer
    type: object
  config.DataSourceConfig:
    properties:
      datasources:
        items:
          $ref: '#/definitions/config.DataSource'
        type: array
      global_options:
        additionalProperties:
          type: integer
        type: object
    type: object
  config.Database:
    properties:
      db_name:
        description: Name of the database
        type: string
      host:
        description: Host of the database
        type: string
      options:
        description: Extra options used while connecting to the database
        type: string
      password:
        description: Password for authentication
        type: string
      port:
        description: Port of the database
        type: string
      primary:
        description: Whether this database is the primary store
        type: boolean
      system:
        description: Database system type (Postgres, MySQL, etc.)
        type: string
      url:
        description: Full URI to the database
        type: string
      username:
        description: Username for authentication
        type: string
    type: object
  config.Scope:
    properties:
      asns:
        description: ASNs specified as in scope
        items:
          type: integer
        type: array
      blacklist:
        description: A blacklist of subdomain names that will not be investigated
        items:
          type: string
        type: array
      cidrs:
        description: CIDR IPNET
        items:
          $ref: '#/definitions/net.IPNet'
        type: array
      domains:
        description: The root domain names that the enumeration will target
        items:
          type: string
        type: array
      ips:
        description: IP Net.IP
        items:
          items:
            format: int32
            type: integer
          type: array
        type: array
      ports:
        description: The ports checked for certificates
        items:
          type: integer
        type: array
    type: object
  config.Transformation:
    properties:
      confidence:
        type: integer
      exclude:
        items:
          type: string
        type: array
      priority:
        type: integer
      ttl:
        type: integer
    type: object
  github_com_owasp-amass_amass_v5_config.Config:
    properties:
      active:
        description: Determines if zone transfers will be attempted
        type: boolean
      alt_worldlist:
        items:
          type: string
        type: array
      alterations:
        description: Will discovered subdomain name alterations be generated?
        type: boolean
      brute_force:
        description: Will the enumeration including brute forcing techniques
        type: boolean
      database:
        description: The graph databases used by the system / enumerations
        items:
          $ref: '#/definitions/config.Database'
        type: array
      datasource_config:
        allOf:
        - $ref: '#/definitions/config.DataSourceConfig'
        description: The data source configurations
      resolvers:
        description: Resolver settings
        items:
          type: string
        type: array
      rigid_boundaries:
        description: Determines rigidness of the enumeration
        type: boolean
      scope:
        allOf:
        - $ref: '#/definitions/config.Scope'
        description: Scope struct that contains ASN, CIDR, Domain, IP, and ports in
          scope
      seed:
        allOf:
        - $ref: '#/definitions/config.Scope'
        description: Seed struct that contains the provided names and CIDRs
      transformations:
        additionalProperties:
          $ref: '#/definitions/config.Transformation'
        description: The Transformations map will contain incoming assets, and what
          handlers should be called.
        type: object
      trusted_resolvers:
        items:
          type: string
        type: array
      wordlist:
        description: The list of words to use when generating names
        items:
          type: string
        type: array
    type: object
  net.IPNet:
    properties:
      ip:
        description: network number
        items:
          format: int32
          type: integer
        type: array
      mask:
        description: network mask
        items:
          format: int32
          type: integer
        type: array
    type: object
...
  server.CreateSessionResponse:
    properties:
      sessionToken:
        type: string
    type: object
  server.ErrorResponse:
    properties:
      code:
        type: integer
      details:
        type: string
      error:
        type: string
    type: object
...
  server.ListSessionsResponse:
    properties:
      sessionTokens:
        items:
          type: string
        type: array
    type: object
...
info:
  contact:
    name: OWASP Amass Project
    url: https://github.com/owasp-amass/amass
  description: |-
    HTTP API for managing Amass Engine sessions and submitting Open Asset Model (OAM) assets.
    Create a session with an engine config, query session stats/scope, and ingest typed OAM assets (single or bulk).
    Includes a WebSocket endpoint for streaming session logs.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  termsOfService: https://owasp.org/www-project-amass/
  title: Amass Engine API (v1)
  version: "1.0"
paths:
...
  /v1/sessions:
    post:
      consumes:
      - application/json
      description: Creates a new Amass engine session using the provided configuration
        JSON.
      parameters:
      - description: Engine configuration
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/github_com_owasp-amass_amass_v5_config.Config'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/server.CreateSessionResponse'
        "400":
          description: Invalid JSON or invalid configuration
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Failed to create session
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Create a new engine session
      tags:
      - sessions
...
  /v1/sessions/list:
    get:
      description: Returns the session tokens for all currently active sessions.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.ListSessionsResponse'
        "404":
          description: Zero sessions found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: List active sessions
      tags:
      - sessions
swagger: "2.0"
...